Added EventListViewController and delete event list view.
